examples/applications/bts/pyproject.toml
[project]
name = "rohe-bts"
version = "0.1.0"
description = "Building energy time-series forecasting pipeline - ROHE reference application"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rohe-sdk",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
]
